pyproject.toml
[project]
name = "wifimon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.2",
    "libpcap>=1.11.0b24",
    "netifaces>=0.11.0",
    "pyx>=0.16",
    "scapy>=2.6.1",
    "sdbus-networkmanager>=2.0.0",
    "vpython>=7.6.5",
]
